pyproject.toml
[project]
name = "rag1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "numpy>=2.3.4",
    "pypdf2>=3.0.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
]
